requirements.txt
asgiref==3.8.1
babel==2.16.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-money==0.7.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
drf-yasg==1.21.7
filetype==1.2.0
inflection==0.5.1
Markdown==3.7
packaging==24.1
pillow==10.4.0
py-moneyed==3.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
django
gunicorn
whitenoise
vercel-python

